+++ HISTORY.md
@@ -4,6 +4,7 @@
 
 ### 2019-01-09
 * Morphic: recognize data sets in dropped text files (csv, json)
+* Lists: updated list documentation, enabled table support by default
 
 ### 2019-01-09
 * Store: tweaked format for serializing atomic data lists